backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-speech==1.42.0
azure-common==1.1.28
azure-core==1.32.0
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isodate==0.7.2
jiter==0.8.2
openai==1.60.2
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
pytest-django==4.10.0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
